requirements.txt
Django==1.9.5
djangorestframework==3.3.3
MySQL-python==1.2.5
